# tests/contracts/fixtures/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ComplianceIQ Core API",
    "description": "\nThe authoritative Core Platform & Data Service for ComplianceIQ.\n\n**Core owns the truth.** Findings, resources, scans and compliance scores\nare produced and owned here. The AI Service reasons *over* this data via\nthis API and never becomes its owner; the dashboard presents it and must\nkeep working when the AI Service is unavailable.\n\n### Authentication\nEvery `/api/v1` endpoint requires `Authorization: Bearer <JWT>` (RS256).\nThe **tenant is taken from the verified token** \u2014 there is no `tenant_id`\nparameter anywhere in this API, and one cannot be supplied.\n\n### Correlation\nSend `X-Correlation-ID` and it is preserved and echoed; omit it and one\nis generated. It appears on every response, including errors.\n\n### Errors\nEvery error \u2014 including validation and unexpected failures \u2014 returns the\nsame envelope: `{\"error\": {\"code\", \"message\", \"correlation_id\", \"details\"}}`.\nBranch on `code`; it is stable. `message` is for humans.\n\n### Versioning\nBreaking changes require a new prefix (`/api/v2`). Additive fields may\nappear in `v1` responses, so clients must ignore unknown fields.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/scans": {
      "post": {
        "tags": [
          "scans"
        ],
        "summary": "Trigger a compliance scan",
        "description": "**Asynchronous.** Returns `202 Accepted` with a `scan_key` and `status: queued`.\n\nThe scan has NOT run when this returns. Poll `GET /api/v1/scans/{scan_key}` until `status` is terminal (`completed`, `partial`, `failed`, `cancelled`).\n\n`partial` means the scan ran but could not enumerate everything \u2014 check `errors`. It is **not** a success: an unreachable service was not verified.",
        "operationId": "submit_scan_api_v1_scans_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanSubmissionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Lacking the 'scanner' role. Triggering a scan is separated from reading data because it spends money and calls real cloud APIs.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request body or parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "A scan for this exact target is already queued or running. Scan keys are deterministic, so a duplicate submission is detected rather than starting a second concurrent scan of the same account.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "tags": [
          "scans"
        ],
        "summary": "List recent scans",
        "operationId": "list_scans_api_v1_scans_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScanResource"
                  },
                  "title": "Response List Scans Api V1 Scans Get"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Lacking the required role.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request body or parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/scans/{scan_key}": {
      "get": {
        "tags": [
          "scans"
        ],
        "summary": "Get scan status and results summary",
        "operationId": "get_scan_api_v1_scans__scan_key__get",
        "parameters": [
          {
            "name": "scan_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Deterministic scan key from the submission response.",
              "title": "Scan Key"
            },
            "description": "Deterministic scan key from the submission response."
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResource"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Lacking the required role.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request body or parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No such scan in this tenant. Identical response whether the scan does not exist or belongs to another tenant.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stable machine-readable error code. Branch on this."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable explanation. May be reworded; do not parse."
          },
          "correlation_id": {
            "type": "string",
            "title": "Correlation Id",
            "description": "Echoes X-Correlation-ID; quote it in bug reports."
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "correlation_id"
        ],
        "title": "ErrorDetail"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "description": "The response body of EVERY error this API returns."
      },
...
      "ScanErrorResource": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "service": {
            "type": "string",
            "title": "Service"
          },
          "operation": {
            "type": "string",
            "title": "Operation"
          },
          "error_code": {
            "type": "string",
            "title": "Error Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "provider",
          "service",
          "operation",
          "error_code",
          "message",
          "retryable",
          "occurred_at"
        ],
        "title": "ScanErrorResource"
      },
      "ScanRequest": {
        "properties": {
          "provider": {
            "type": "string",
            "enum": [
              "aws",
              "azure"
            ],
            "title": "Provider",
            "description": "Cloud provider to scan. GCP is not implemented; see docs/architecture/phase-5-audit.md."
          },
          "account_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id",
            "description": "AWS account id or Azure subscription id."
          },
          "directory_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Directory Id",
            "description": "Azure tenant (directory) id, where applicable."
          },
          "regions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Regions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "provider"
        ],
        "title": "ScanRequest",
        "description": "``POST /api/v1/scans`` body.\n\nNo ``tenant_id``: it comes from the verified token. Accepting one\nhere would create exactly the \"trust the client's tenant\" hole \u00a712\nforbids, and ``extra=\"forbid\"`` means sending one is a 422 rather\nthan a silently ignored field."
      },
      "ScanResource": {
        "properties": {
          "scan_key": {
            "type": "string",
            "title": "Scan Key"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "account_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id"
          },
          "directory_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Directory Id"
          },
          "regions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Regions"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "partial",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "resource_count": {
            "type": "integer",
            "title": "Resource Count"
          },
          "finding_count": {
            "type": "integer",
            "title": "Finding Count"
          },
          "pass_count": {
            "type": "integer",
            "title": "Pass Count"
          },
          "fail_count": {
            "type": "integer",
            "title": "Fail Count"
          },
          "indeterminate_count": {
            "type": "integer",
            "title": "Indeterminate Count"
          },
          "critical_count": {
            "type": "integer",
            "title": "Critical Count"
          },
          "high_count": {
            "type": "integer",
            "title": "High Count"
          },
          "medium_count": {
            "type": "integer",
            "title": "Medium Count"
          },
          "low_count": {
            "type": "integer",
            "title": "Low Count"
          },
          "error_count": {
            "type": "integer",
            "title": "Error Count"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ScanErrorResource"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "scan_key",
          "tenant_id",
          "provider",
          "account_id",
          "directory_id",
          "regions",
          "status",
          "started_at",
          "completed_at",
          "duration_seconds",
          "resource_count",
          "finding_count",
          "pass_count",
          "fail_count",
          "indeterminate_count",
          "critical_count",
          "high_count",
          "medium_count",
          "low_count",
          "error_count"
        ],
        "title": "ScanResource",
        "description": "A scan's full state.\n\n``partial`` is a first-class status, not a variant of ``completed``:\na scan that enumerated S3 but was denied KMS has NOT verified KMS,\nand reporting it as completed would tell an auditor that KMS was\nchecked and found compliant. Consult ``errors`` whenever status is\n``partial``."
      },
      "ScanSubmissionResponse": {
        "properties": {
          "scan_key": {
            "type": "string",
            "title": "Scan Key"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "partial",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Submitted At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "scan_key",
          "status",
          "tenant_id",
          "submitted_at"
        ],
        "title": "ScanSubmissionResponse",
        "description": "``202 Accepted``. An id and a status \u2014 never results.\n\nA submission response containing findings would imply the scan had\nfinished, which is the misconception \u00a726 asks the contract to avoid.\nPoll ``GET /api/v1/scans/{scan_key}`` for progress."
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "RS256 JWT issued by complianceiq-core. Required claims: sub, tenant_id, roles, iss=complianceiq-core, aud=complianceiq, exp. Verify against /.well-known/jwks.json."
      }
    }
  }
}